ImageInpaintingBackend/requirements.txt
django
django-rest-framework
torch
torchvision
tensorflow
sklearn
tqdm
pyyaml
opencv-python